# Conflicts:
#	src/test/java/chatApp/controller/AuthControllerTest.java

+++ src/main/java/chatApp/service/UserService.java
@@ -72,10 +72,10 @@ public class UserService {
                 dbUser.setDateOfBirth(user.getDateOfBirth());
                 dbUser.setAge(calcAge(user.getDateOfBirth()));
             }
-            if (!user.getPhoto().equals("")) {
+            if (user.getPhoto() != null && !user.getPhoto().equals("")) {
                 dbUser.setPhoto(user.getPhoto());
             }
-            if (!user.getDescription().equals("")) {
+            if (user.getDescription() != null && !user.getDescription().equals("")) {
                 dbUser.setDescription(user.getDescription());
             }
             logger.info("Update the user, and save updating in DB");
@@ -97,9 +97,17 @@ public class UserService {
         try {
             logger.info("Delete the user token");
             String userEmail = authService.getKeyTokensValEmails().get(token);
+            if (userRepository.findByEmail(userEmail) == null) {
+                logger.error(emailNotExistsMessage);
+                throw new IllegalArgumentException(emailNotExistsMessage);
+            }
             authService.getKeyTokensValEmails().remove(token);
             authService.getKeyEmailsValTokens().remove(userEmail);
             User dbUser = User.dbUser(userRepository.findByEmail(userEmail));
+            if (dbUser == null) {
+                logger.error(userNotFound);
+                throw new IllegalArgumentException(userNotFound);
+            }
             logger.info("If the user guest delete him from DB else update his status to offline");
             if (dbUser.getType().equals(UserType.GUEST) && dbUser.getEmail().contains("chatappsystem")) {
                 userRepository.delete(dbUser);
@@ -107,7 +115,7 @@ public class UserService {
             }
             dbUser.setUserStatus(UserStatuses.OFFLINE);
             return userRepository.save(dbUser);
-        } catch (JpaSystemException e) {
+        } catch (IllegalArgumentException | JpaSystemException e) {
             throw new IllegalArgumentException(e);
         }
     }
@@ -126,6 +134,10 @@ public class UserService {
             if (userEmail == null) {
                 logger.error(tokenSessionExpired);
                 throw new IllegalArgumentException(tokenSessionExpired);
+            }
+            if (userRepository.findByEmail(userEmail).getType() != UserType.ADMIN) {
+                logger.error(notAdminUser);
+                throw new IllegalArgumentException(notAdminUser);
             }
             if (userRepository.getById(id) == null) {
                 throw new IllegalArgumentException(emailNotExistsMessage(userEmail));
